feat: formulario seleccion de directoras

+++ source-code/onboarding/model/src/main/java/bo/com/mc4/onboarding/model/business/dto/FrmDireccionDto.java
@@ -11,8 +11,8 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class FrmDireccionDto {
-    private Long departamentoId;
-    private Long localidadId;
+    private Integer departamentoId;
+    private Integer localidadId;
     private String barrioZona;
     private String avenidaCalle;
     private String nroCasa;
